A bit more refactoring of `rsmtool.utils` package
